implement community page
