+++ Frontend/src/components/Sidebar/index.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect }  from 'react';
+import { useState, useEffect } from 'react';
 import LiveHelpIcon from '@mui/icons-material/LiveHelp';
 import MenuBookIcon from '@mui/icons-material/MenuBook';
 import MapIcon from '@mui/icons-material/Map';
@@ -26,12 +26,10 @@ function indexIcon(index) {
         />
       );
     case 1:
-      return (<TaskIcon className={styles.coloredIcon} />);
-    case 2:
       return (<AccountBalanceIcon className={styles.coloredIcon} />);
-    case 3:
+    case 2:
       return (<SearchIcon className={styles.coloredIcon} />);
-    case 4:
+    case 3:
       return (<MenuBookIcon className={styles.coloredIcon} />);
     default:
       return (<LiveHelpIcon className={styles.coloredIcon} />);
@@ -40,18 +38,18 @@ function indexIcon(index) {
 
 export default function Sidebar(props) {
   const { setUser } = props;
-  const routes = ['genemapper', 'teams', 'institutions', 'search/teams', 'documentation', 'help'];
-  const titles = ['Gene Mapper', 'Teams', 'Institutions', 'Search', 'Documentation', 'Help'];
+  const routes = ['genemapper', 'community', 'search/teams', 'documentation', 'help'];
+  const titles = ['Gene Mapper', 'Community', 'Search', 'Documentation', 'Help'];
   const { url } = useRouteMatch();
   const location = useLocation();
   const path = location.pathname;
   const settingsPath = '/sequencer/settings';
   const [checkPathActive, setCheckPathActive] = useState("/")
   const pathRegex = new RegExp(".*\/sequencer\/(\\w+)(?=\/|)", "g")
-  
+
   useEffect(() => {
-    const result = pathRegex.exec(path) 
-    if(result) {
+    const result = pathRegex.exec(path)
+    if (result) {
       setCheckPathActive(result[1])
     }
   }, [path])
