+++ Frontend/src/components/teams/overview/TeamOverview/teamOverview.module.css
@@ -1,0 +1,3 @@
+.cardSpacing {
+  margin-bottom: 16px;
+}
