+++ Frontend/src/views/CommunityOverview/communityOverview.module.css
@@ -1,0 +1,3 @@
+.content {
+  padding-top: 16px;
+}
