+++ Frontend/src/views/GeneMapper/Home/index.jsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box';
 import {
   Typography, createTheme, ThemeProvider, Stack, TextField,
 } from '@mui/material';
-import PlusIcon from 'components/GeneMapper/plusIcon';
+import PlusIcon from 'components/general/PlusIcon';
 import ProjectBarCard from 'components/GeneMapper/projectBarCard';
 import SearchIcon from '@mui/icons-material/Search';
 import ProjectService from 'shared/services/Project.service';
@@ -13,6 +13,7 @@ import ProjectMock from 'shared/services/mock/projects';
 import AtlasService from 'shared/services/Atlas.service';
 import ModelService from 'shared/services/Model.service';
 import TeamService from 'shared/services/Team.service';
+import { useHistory } from 'react-router-dom';
 
 const theme = createTheme({
   palette: {
@@ -25,13 +26,6 @@ const theme = createTheme({
   },
 });
 
-const themeIcon = createTheme({
-  palette: {
-    primary: {
-      main: '#5676E4',
-    },
-  },
-});
 
 function GeneMapperHome() {
   const [projects, setProjects] = useState([]);
@@ -41,6 +35,8 @@ function GeneMapperHome() {
 
   const [findString, setFindString] = useState('');
   const [submissionProgress, setSubmissionProgress] = useSubmissionProgress();
+
+  const history = useHistory();
 
   const handleDeleteItem = (id) => {
     setProjects(projects.filter((object) => object._id != id));
@@ -107,9 +103,7 @@ function GeneMapperHome() {
           <Stack direction="row" className="stack" alignItems="Center">
 
             <Typography variant="h5" sx={{ pr: 1 }}>Your Mappings</Typography>
-            <ThemeProvider theme={themeIcon}>
-              <PlusIcon />
-            </ThemeProvider>
+            <PlusIcon onClick={() => { history.push('genemapper/create'); }} />
           </Stack>
           <TextField
             id="outlined-basic"
@@ -118,7 +112,7 @@ function GeneMapperHome() {
                 <SearchIcon />
                 Find a Mapping
               </Stack>
-        )}
+            )}
             variant="outlined"
             size="small"
             value={findString}
@@ -144,7 +138,7 @@ function GeneMapperHome() {
                   submissionProgress={submissionProgress.uploadId === project.uploadId
                     ? submissionProgress : null}
                   setSubmissionProgress={submissionProgress.uploadId === project.uploadId
-                    ? setSubmissionProgress : () => {}}
+                    ? setSubmissionProgress : () => { }}
                 />
 
               );
