--- Frontend/src/views/InstitutionOverview/institutionOverview.module.css
@@ -1,7 +1,0 @@
-.content {
-    padding-top: 16px;
-}
-
-.cardSpacing {
-    margin-bottom: 16px;
-}
